third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-uncaught-promise-on-pause-expected.txt

CONSOLE ERROR: line 16: Unhandled promise rejection
CONSOLE ERROR: line 34: Unhandled promise rejection
Tests uncaught promise rejections fired during pause.

Set timer for test function.
debugger-uncaught-promise-on-pause.html:10 Console was cleared

    debugger-uncaught-promise-on-pause.html:16 Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: Error: inspector.err1}
        __proto__: Promise
        [[PromiseStatus]]: "rejected"
        [[PromiseValue]]: Error: inspector.err1
        debugger-uncaught-promise-on-pause.html:16 runPromises
        debugger-uncaught-promise-on-pause.html:41 timeout

    debugger-uncaught-promise-on-pause.html:34 Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: Error: inspector.err2}
        __proto__: Promise
        [[PromiseStatus]]: "rejected"
        [[PromiseValue]]: Error: inspector.err2
        debugger-uncaught-promise-on-pause.html:34 runPromises
        debugger-uncaught-promise-on-pause.html:41 timeout
